+++ src/internal/persistence/models/store.go
@@ -7,8 +7,8 @@ type Store struct {
 	ClientID      int64     `gorm:"column:ClientID;not null"`
 	StoreName     string    `gorm:"column:StoreName;type:varchar(150);not null"`
 	Address       string    `gorm:"column:Address;type:varchar(150);not null"`
-	CityID        int8      `gorm:"column:CityID;not null"`
-	StateID       int8      `gorm:"column:StateID;not null"`
+	CityID        int16     `gorm:"column:CityID;not null"`
+	StateID       int16     `gorm:"column:StateID;not null"`
 	Pincode       string    `gorm:"column:Pincode;type:varchar(6);not null"`
 	ContactNumber string    `gorm:"column:ContactNumber;type:varchar(10);not null"`
 	EmailID       string    `gorm:"column:EmailID;type:varchar(75);not null"`
